+++ hardhat.config.ts
@@ -14,6 +14,7 @@ import {
   HardhatNetworkUserConfig,
   NetworkUserConfig,
 } from 'hardhat/types'
+import { EthGasReporterConfig, RemoteContract } from "hardhat-gas-reporter/src/types"
 import path from 'path'
 
 config()
@@ -96,6 +97,39 @@ const networkConfig = (config: NetworkUserConfig): NetworkUserConfig => {
 }
 
 
+const getRemoteContracts = (networkType:string) : RemoteContract[] => {
+
+  const remoteContracts:RemoteContract[] = []
+  
+  const preCompilesPath = `deployments/${networkType.toLowerCase()}`
+  for (const fileName of fs.readdirSync(preCompilesPath)){
+    const artifactPath = path.join(process.cwd(), preCompilesPath, fileName) 
+     
+    if(fileName.startsWith('.') || !fileName.endsWith('.json'))continue
+
+    try{ 
+
+    const preDeployed =  JSON.parse( fs.readFileSync(artifactPath, 'utf8' ) )
+     
+    if (preDeployed.address){
+      remoteContracts.push({
+        name: preDeployed.artifactName,
+        abi: preDeployed.abi,
+        address: preDeployed.address
+      })
+    } 
+
+    }catch(e){
+      console.error('Could not parse artifact file: ',e)
+    }
+
+  //console.log('preDeployed',preDeployed.artifactName)
+
+  }
+
+  return remoteContracts
+}
+
 
 
 
@@ -147,7 +181,8 @@ export default <HardhatUserConfig>{
     noColors: !!SAVE_GAS_REPORT,
     showMethodSig: false,
     showTimeSpent: true,
-    onlyCalledMethods: false
+    onlyCalledMethods: false,
+    remoteContracts: getRemoteContracts('hardhat')
   },
   namedAccounts: {
     deployer: '0xAFe87013dc96edE1E116a288D80FcaA0eFFE5fe5',
